Move dag_edges and task_group_to_dict to corresponding util modules
The methods were implemented in "view" but they were used in other
places (in dot_renderer) so they conceptually belong to common
util code. Having those in a wrong package (airflow/www) caused
the tests to pass because selective checks did nor realise that
change in "airflow/www" also requires running Core tests.

This PR moves the methods to "airflow/utils". The methods
are imported in the "views" module so even if someone used them
from there, they will stil be available there, so the change is
fully backwards compatible (even if those are not "public"
airflow API methods.

Follow up after #26188
